Address code review nitpick
Also renamed the method to make more sense.
Though honestly the entirety of this project is a bit untested, I'd not be surprised
if the method doesn't even correctly work. But I'd first want to get micromixin-remapper
developed and then we can debug both projects at the same time within gslStarplane.