Fix indys getting incorrectly remapped
The lambda's name is stored as the indy's name, but was previously completely ignored
by stianloader-remapped. However, the JVM wouldn't turn a blind eye to this error and
would cause a crash whenever the indy is used.

It is possible that this causes regressions with non-lambda indys. We shall see about
that in the future.